# Conflicts:
#	app/src/main/java/com/hackathon/alddeul_babsang/presentation/report/screen/ReportScreen.kt
#	app/src/main/java/com/hackathon/alddeul_babsang/presentation/report/screen/ReportViewModel.kt

+++ app/src/main/java/com/hackathon/alddeul_babsang/data/datasourceimpl/ProfileDataSourceImpl.kt
@@ -11,7 +11,7 @@ class ProfileDataSourceImpl @Inject constructor(
     private val profileApiService: ProfileApiService
 ) : ProfileDataSource {
     override suspend fun getLikes(
-        userId: Long
+        userId: Int
     ): BaseResponse<ResponseLikesDto> {
         return profileApiService.getLikes(userId)
     }
